server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
googlemaps==4.10.0
pydantic==2.3.0
requests==2.31.0